requirements.txt
tensorflow==2.18.0
numpy==2.0.2
streamlit==1.43.2
pillow==11.1.0